close issues/604
